+++ poulpy-core/src/lib.rs
@@ -78,7 +78,7 @@ pub mod layouts;
 pub use api::*;
 pub use default::encryption::*;
 pub use default::linear_transformation::*;
-pub(crate) use default::noise::{log2_std_noise_glwe_tensor, var_noise_gglwe_product_v2};
+pub(crate) use default::noise::log2_std_noise_glwe_tensor;
 pub use default::operations::*;
 pub use default::polynomial_evaluation::{BSGSOps, GiantStepTensorBounds};
 pub use dist::*;