+++ poulpy-core/src/test_suite/automorphism/ggsw_ct.rs
@@ -14,7 +14,7 @@ use crate::{
         GLWEAutomorphismKeyPreparedFactory, GLWEInfos, GLWESecret, GLWESecretPreparedFactory, ModuleCoreAlloc,
         prepared::{GGLWEToGGSWKeyPrepared, GLWEAutomorphismKeyPrepared, GLWESecretPrepared},
     },
-    noise::noise_ggsw_keyswitch,
+    noise::GGLWENoiseModel,
     test_suite::{
         download_scalar_znx, upload_gglwe_to_ggsw_key, upload_ggsw, upload_glwe_automorphism_key, upload_glwe_secret,
         upload_scalar_znx,
@@ -47,9 +47,6 @@ where
     let p: i64 = -5;
     for rank in 1_usize..3 {
         for dsize in 1..max_dsize + 1 {
-            let k_ksk: usize = k_in + key_base2k * dsize;
-            let k_tsk: usize = k_ksk;
-
             let n: usize = module.n();
             let dnum_in: usize = k_in / in_base2k;
             let dnum_ksk: usize = k_in.div_ceil(key_base2k * dsize);
@@ -177,18 +174,14 @@ where
             let pt_scalar_noise = download_scalar_znx::<BE>(&pt_scalar_backend);
 
             let max_noise = |col_j: usize| -> f64 {
-                noise_ggsw_keyswitch(
-                    n as f64,
-                    key_base2k * dsize,
+                auto_key_layout.log2_std_noise_ggsw_keyswitch(
+                    &tsk_layout,
                     col_j,
+                    &ggsw_in_layout,
                     var_xs,
-                    0f64,
+                    DEFAULT_SIGMA_XE * DEFAULT_SIGMA_XE,
                     DEFAULT_SIGMA_XE * DEFAULT_SIGMA_XE,
                     0f64,
-                    rank as f64,
-                    k_in,
-                    k_ksk,
-                    k_tsk,
                 ) + 0.5
             };
 
@@ -241,9 +234,6 @@ where
     let p: i64 = -1;
     for rank in 1_usize..3 {
         for dsize in 1..max_dsize + 1 {
-            let k_ksk: usize = k_out + key_base2k * dsize;
-            let k_tsk: usize = k_ksk;
-
             let n: usize = module.n();
             let dnum_in: usize = k_out / out_base2k;
             let dnum_ksk: usize = k_out.div_ceil(key_base2k * dsize);
@@ -358,18 +348,14 @@ where
             let pt_scalar_noise = download_scalar_znx::<BE>(&pt_scalar_backend);
 
             let max_noise = |col_j: usize| -> f64 {
-                noise_ggsw_keyswitch(
-                    n as f64,
-                    key_base2k * dsize,
+                auto_key_layout.log2_std_noise_ggsw_keyswitch(
+                    &tsk_layout,
                     col_j,
+                    &ggsw_out_layout,
                     var_xs,
-                    0f64,
+                    DEFAULT_SIGMA_XE * DEFAULT_SIGMA_XE,
                     DEFAULT_SIGMA_XE * DEFAULT_SIGMA_XE,
                     0f64,
-                    rank as f64,
-                    k_out,
-                    k_ksk,
-                    k_tsk,
                 ) + 4.0
             };