+++ poulpy-core/src/test_suite/external_product/ggsw_ct.rs
@@ -13,7 +13,7 @@ use crate::{
         GGSW, GGSWInfos, GGSWLayout, GGSWPreparedFactory, GLWEInfos, GLWESecret, GLWESecretPreparedFactory, ModuleCoreAlloc,
         prepared::{GGSWPrepared, GLWESecretPrepared},
     },
-    noise::noise_ggsw_product,
+    noise::GGSWNoiseModel,
 };
 
 #[allow(clippy::too_many_arguments)]
@@ -39,8 +39,6 @@ where
 
     for rank in 1_usize..3 {
         for dsize in 1..max_dsize + 1 {
-            let k_apply: usize = k_in + key_base2k * dsize;
-
             let n: usize = module.n();
             let dnum: usize = k_in.div_ceil(key_base2k * dsize);
             let dnum_in: usize = k_in / in_base2k;
@@ -141,18 +139,14 @@ where
             let var_a1_err: f64 = 1f64 / 12f64;
 
             let max_noise = |_col_j: usize| -> f64 {
-                noise_ggsw_product(
-                    n as f64,
-                    key_base2k * dsize,
+                ggsw_apply_infos.log2_std_noise_external_product(
+                    &ggsw_in_infos,
                     0.5,
                     var_msg,
                     var_a0_err,
                     var_a1_err,
                     var_gct_err_lhs,
                     var_gct_err_rhs,
-                    rank as f64,
-                    k_in,
-                    k_apply,
                 ) + 0.5
             };
 
@@ -201,8 +195,6 @@ where
 
     for rank in 1_usize..3 {
         for dsize in 1..max_dsize + 1 {
-            let k_apply: usize = k_out + key_base2k * dsize;
-
             let n: usize = module.n();
             let dnum: usize = k_out.div_ceil(dsize * key_base2k);
             let dnum_in: usize = k_out / out_base2k;
@@ -294,18 +286,14 @@ where
             let var_a1_err: f64 = 1f64 / 12f64;
 
             let max_noise = |_col_j: usize| -> f64 {
-                noise_ggsw_product(
-                    n as f64,
-                    key_base2k * dsize,
+                ggsw_apply_infos.log2_std_noise_external_product(
+                    &ggsw_out_infos,
                     0.5,
                     var_msg,
                     var_a0_err,
                     var_a1_err,
                     var_gct_err_lhs,
                     var_gct_err_rhs,
-                    rank as f64,
-                    k_out,
-                    k_apply,
                 ) + 0.5
             };